requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.2.3
djangorestframework==3.12.4
gunicorn==20.1.0
idna==2.10
MarkupSafe==2.0.1
packaging==20.9
psycopg2==2.8.6
pyparsing==2.4.7
python-decouple==3.4
pytz==2021.1
requests==2.25.1
simplejson==3.17.2
sqlparse==0.4.1
urllib3==1.26.5
whitenoise==5.2.0
